requirements.txt
beautifulsoup4==4.11.1
python-decouple==3.6
requests==2.24.0
